Editar/Excluir/Criar Pontuação

+++ website/views.py
@@ -9,6 +9,7 @@ from .models import *
 from .forms import *
 
 def home(request):
+	print(str(request.user.user_permissions))
 	return render(request, 'home.html')
 
 def login(request):
@@ -221,6 +222,8 @@ def match_details(request, match_id):
 @login_required()
 def edit_match(request, match_id):
 	match = get_object_or_404(Match, id=match_id)
+	if request.user.username!=match.responsible.username:
+		return redirect('/')
 	if request.method == 'POST':
 		form = MatchForm(request.POST, instance=match)
 		if form.is_valid():
@@ -228,6 +231,7 @@ def edit_match(request, match_id):
 			return redirect('/jogos/' + match_id)
 	else:
 		form = MatchForm(instance=match)
+
 
 	competition = get_object_or_404(Competition, id=match.competition.id)
 	context = {
@@ -252,6 +256,8 @@ def edit_match(request, match_id):
 @login_required()
 def remove_match(request, match_id):
 	match = get_object_or_404(Match, id=match_id)
+	if request.user.username!=match.responsible.username:
+		return redirect('/')
 	competition_id = match.competition.id
 	match.delete()
 	return redirect('/competicoes/' + str(competition_id))
@@ -297,7 +303,7 @@ def attend_to_match(request, match_id):
 @login_required()
 def leave_match(request, match_id):
 	match = get_object_or_404(Match, id=match_id)
-	match.participants.remove(request.user)
+	match.teams.remove(request.user)
 	score = MatchScore.objects.get(match=match, participant=request.user)
 	score.delete()
 	return redirect('/jogos/' + str(match_id))
@@ -337,10 +343,10 @@ def list_results(request):
 		not_matchs = Match.match_not_ready(request) + Match.matchs_ready_to_publish_result(request)
 	for match in matchs:
 			matchScore = MatchScore.objects.all().filter(match=match)
-			if match.participants.count()!=matchScore.count():
+			if match.teams.count()!=matchScore.count():
 				match.first_place=None
 				match.save()
-			elif match.participants.count()==matchScore.count() and match.first_place:
+			elif match.teams.count()==matchScore.count() and match.first_place:
 				boolean = True
 	context = {
 		'title': 'Resultados',
@@ -361,12 +367,14 @@ def list_incomplete_or_not_plubished_results(request, user_name):
 		return redirect('/resultados')
 	MATCHS = Match.matchs_ready_to_publish_result(request)
 	MATCH  = Match.match_not_ready(request)
+	MATC   = Match.match_already_published(request)
 	context = {
-		'title': 'Partidas - '+ user_name,
+		'title': 'Partidas - '+ request.user.get_full_name(),
 		'USER': request.user.username,
 		'USER_MORE': user_name,
 		'matchs': MATCHS,
 		'matchs_not_ready': MATCH,
+		'match_already_published': MATC,
 		'breadcrumb': [
 			{'name': 'Início', 'link': '/'},
 			{'name': 'Torneios', 'link': '/torneios'},
@@ -380,8 +388,10 @@ def list_incomplete_or_not_plubished_results(request, user_name):
 
 def match_score(request, user_name, match_id):
 	match = get_object_or_404(Match, id=match_id)
+	if user_name!=match.responsible.username:
+		return redirect("/resultados")
 	matchScore = MatchScore.objects.all().filter(match=match)
-	if not matchScore.count() == match.participants.count() or not match.first_place:
+	if not matchScore.count() == match.teams.count() or not match.first_place:
 		if request.user.is_anonymous():
 			return redirect('/resultados')
 
@@ -410,11 +420,9 @@ def publish_result(request, user_name, match_id):
 	if request.user.username != user_name:
 		return redirect('/resultados')
 	match = get_object_or_404(Match, id=match_id)
-						# Pega todos os MatchScore da partida, ordena de maneira 
-						# decrescente e pega o primeiro (Maior pontuação)
-	match.first_place = MatchScore.objects.all().filter(match=match).order_by('-score').first().participant
+	match.first_place = MatchScore.objects.all().filter(match=match).order_by('-score').first().team
 	match.save()
-	return redirect('/resultados/'+ user_name + '/' +match_id)
+	return redirect('/resultados/'+user_name)
 
 @login_required
 def list_incomplete_scores(request, user_name, match_id):
@@ -422,18 +430,12 @@ def list_incomplete_scores(request, user_name, match_id):
 		return redirect('/resultados')
 	match = get_object_or_404(Match, id=match_id)
 	scores = MatchScore.objects.all().filter(match = match)
-	participants_complete_score = []
-	participants_incomplete_score = []
+	teams_incomplete_score = []
 	
-	for score in scores:                        #Percorre MatchScore já cadastrados
-		for participant in Participant.objects.all():  #Percorre Participantes da match
-			if score.participant == participant:#Se Participante == MatchScore.participante 
-				participants_complete_score.append(participant)		#Adiciona a matchScore na lista de MS completa
-
-	for participant in match.participants.all():
-		matchScores = scores.filter(participant=participant)
+	for team in match.teams.all():
+		matchScores = scores.filter(team=team)
 		if not matchScores:
-			participants_incomplete_score.append(participant)
+			teams_incomplete_score.append(team)
 	
 	context = {
 		'title': 'Pontuação Individual - '+ str(match),
@@ -441,7 +443,7 @@ def list_incomplete_scores(request, user_name, match_id):
 		'USER_MORE': user_name,
 		'match': match,
 		'matchScores': scores.order_by('-score'),
-		'participants_without_matchScore': participants_incomplete_score,
+		'teams_without_matchScore': teams_incomplete_score,
 
 		'breadcrumb': [
 			{'name': 'Início', 'link': '/'},
@@ -456,11 +458,11 @@ def list_incomplete_scores(request, user_name, match_id):
 	return render(request, 'matchscore-add.html', context)
 
 @login_required
-def add_matchScore(request, user_name, match_id, participant_id):
+def add_matchScore(request, user_name, match_id, team_id):
 	if request.user.username != user_name:
 		return redirect('/resultados')
 	match = get_object_or_404(Match, id=match_id)
-	participant = get_object_or_404(Participant, id=participant_id)
+	team = get_object_or_404(Team, id=team_id)
 	competition = match.competition
 	tournament = competition.tournament
 
@@ -468,18 +470,20 @@ def add_matchScore(request, user_name, match_id, participant_id):
 		form = MatchScoreForm(request.POST)
 		if form.is_valid():
 			matchScore = form.save(commit=False)
+			team.score = matchScore.score
+			team.save()
 			match.first_place=None
 			match.save()
 			matchScore.match=match
-			matchScore.participant=participant
+			matchScore.team=team
 			matchScore.save()
 			return redirect("/pontuacao/" + user_name + '/' + str(match_id) + '/')
 	else:
 		form = MatchScoreForm()
 
 	context = {
-		'title': "Pontuação" + str(match) + ' - ' + participant.name,
-		'action': '/pontuacao/' + user_name + '/' + str(match_id) + '/' + str(participant_id) + '/',
+		'title': "Pontuação - " + str(match) + ' - ' + team.name,
+		'action': '/pontuacao/' + user_name + '/' + str(match_id) + '/' + str(team_id) + '/',
 		'cancel': '/pontuacao/' + user_name + '/' + str(match_id),
 		'form': form,
 		'competition' : competition,
@@ -495,3 +499,39 @@ def add_matchScore(request, user_name, match_id, participant_id):
 	}
 
 	return render(request, 'form.html', context)
+
+@login_required
+def edit_matchScore(request, user_name, matchScore_id):
+	if request.user.username != user_name:
+		return redirect("/resultados")
+	matchScore = get_object_or_404(MatchScore, id=matchScore_id)
+
+	if request.method == 'POST':
+		form = MatchScoreForm(request.POST, instance=matchScore)
+		if form.is_valid():
+			matchScore.match.first_place=None
+			matchScore.match.save()
+			form.save()
+			return redirect('/pontuacao/' + user_name + '/' + str(matchScore.match.id))
+	else:
+		form = MatchScoreForm(instance=matchScore)
+
+	context = {
+		'title': "Pontuação - " + str(matchScore.match) + ' - ' + matchScore.team.name,
+		'action': '/pontuacao/editar/' + user_name + '/' + str(matchScore_id) + '/',
+		'cancel': '/pontuacao/' + user_name + '/' + str(matchScore.match.id),
+		'form': form,
+		'breadcrumb': [
+		]
+	}
+
+	return render(request, 'form.html', context)
+
+@login_required
+def remove_matchScore(request, user_name,matchScore_id):
+	if request.user.username!=user_name:
+		return redirect('/')
+	matchScore = get_object_or_404(MatchScore, id=matchScore_id)
+	match_id = matchScore.match.id
+	matchScore.delete()
+	return redirect('/pontuacao/' + user_name + '/' + str(match_id))
